refactor(consts): separate global constants to ns_consts.py
