+++ scripts/ns_download_models.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python3
 
 import stanza
-from neosca import STANZA_MODEL_DIR
+from neosca.ns_consts import STANZA_MODEL_DIR
 
 stanza.download("en", model_dir=str(STANZA_MODEL_DIR))
