+++ scripts/ns_packaging.spec
@@ -9,7 +9,7 @@ import PyInstaller
 sys.path.insert(0, str(Path(".").parent.absolute() / "src"))
 from neosca.ns_platform_info import IS_MAC, IS_LINUX, IS_WINDOWS  # noqa: I001, E402
 from neosca.ns_about import __title__, __version__  # noqa: I001, E402
-from neosca import ICON_PATH, ICON_MAC_PATH  # noqa: I001, E402
+from neosca.ns_consts import ICON_PATH, ICON_MAC_PATH  # noqa: I001, E402
 
 # https://pyinstaller.org/en/stable/spec-files.html#spec-file-options-for-a-macos-bundle
