+++ src/neosca/ns_lca/ns_lca_counter.py
@@ -11,7 +11,7 @@ from collections.abc import Sequence
 from math import sqrt as _sqrt
 from typing import Literal
 
-from neosca import DATA_DIR
+from neosca.ns_consts import DATA_DIR
 from neosca.ns_io import Ns_IO
 from neosca.ns_lca import word_classifiers
 from neosca.ns_utils import chunks, safe_div, safe_log
