+++ src/neosca/ns_settings/ns_settings.py
@@ -4,7 +4,7 @@ from typing import Any
 
 from PyQt5.QtCore import QSettings
 
-from neosca import SETTING_PATH
+from neosca.ns_consts import SETTING_PATH
 from neosca.ns_settings.ns_settings_default import settings_default
